+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Unreleased
  - Moved `MenuListItem` and `Marker` to `items`
  - Renamed `Select` to `MenuItem`
  - `MenuListItem::draw_styled` and `MenuListItem::set_style` now only takes a font style instead of the whole menu style
+ - Renamed `MenuBuilder::add_item(s)` to `add_menu_item(s)`
 
 ## Removed